Small refactor and UI improvement
